Convert the telegram bot app to Spring based. It allows for easier customization of future work.
