+++ pom.xml
@@ -11,7 +11,6 @@
     <modules>
         <module>covid19-models</module>
         <module>covid19-patients-analyzer</module>
-        <module>covid19-telegram-bot</module>
         <module>covid19-stats</module>
     </modules>
 </project>
